Use a generic sqs queue
